[FEAT] Add new behavior under Retrofit service
  - Add new behavior under Retrofit service to request and receive necessary data for newly added usecase.
  - Corresponding function to initiate new behaviors has been declared.

  - Relocate file from directory model to data.
