+++ app/src/main/java/com/doyoonkim/knutice/data/NoticeRemoteSource.kt
@@ -1,0 +1,57 @@
+package com.doyoonkim.knutice.data
+
+import android.util.Log
+import com.doyoonkim.knutice.model.NoticeCategory
+import com.doyoonkim.knutice.model.NoticesPerPage
+import com.doyoonkim.knutice.model.TopThreeNotices
+import com.example.knutice.BuildConfig
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import retrofit2.http.GET
+import retrofit2.http.Query
+import javax.inject.Inject
+
+class NoticeRemoteSource @Inject constructor() {
+    val retrofit: Retrofit = Retrofit.Builder()
+        .baseUrl(BuildConfig.API_ROOT)
+        .addConverterFactory(GsonConverterFactory.create())
+        .build()
+
+    suspend fun getTopThreeNotice(): TopThreeNotices {
+        Log.d("NoticeRemoteSource", "Start retrofit service")
+        return retrofit.create(KnuticeService::class.java).run {
+            this.getTopThreeNotice()
+        }
+    }
+
+
+    suspend fun getNoticeListPerPage(category: NoticeCategory, lastNttId: Int): NoticesPerPage {
+        Log.d("NoticeRemoteSource", "Start retrofit service")
+        return retrofit.create(KnuticeService::class.java).run {
+            if (lastNttId == 0) {
+                this.getFirstPageOfNotice(category)
+            } else {
+                this.getNoticeListPerPage(category, lastNttId)
+            }
+        }
+    }
+
+}
+
+interface KnuticeService {
+
+    @GET("/open-api/notice")
+    suspend fun getTopThreeNotice(): TopThreeNotices
+
+    @GET("/open-api/notice/list")
+    suspend fun getNoticeListPerPage(
+        @Query("noticeName") category: NoticeCategory,
+        @Query("nttId") lastNttId: Int
+    ): NoticesPerPage
+
+    @GET("/open-api/notice/list")
+    suspend fun getFirstPageOfNotice(
+        @Query("noticeName") category: NoticeCategory
+    ): NoticesPerPage
+
+}
